pyproject.toml
[project]
name = "audio-script"
version = "0.1.0"
description = ""
authors = [
    {name = "Romankk03",email = "126852895+Romankk03@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy (>=2.3.5,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg (>=3.3.1,<4.0.0)",
    "fastapi (>=0.123.5,<0.124.0)",
    "unicorn (>=2.1.4,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "sqlmodel (>=0.0.27,<0.0.28)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
